Run eject to rename app to PostAR
