+++ android/app/build.gradle
@@ -87,7 +87,7 @@ android {
     buildToolsVersion "23.0.1"
 
     defaultConfig {
-        applicationId "com.mobileapp"
+        applicationId "com.postar"
         minSdkVersion 16
         targetSdkVersion 22
         versionCode 1
@@ -127,7 +127,6 @@ android {
 
 dependencies {
     compile project(':react-native-sensors')
-    compile project(':react-native-camera')
     compile fileTree(dir: "libs", include: ["*.jar"])
     compile "com.android.support:appcompat-v7:23.0.1"
     compile "com.facebook.react:react-native:+"  // From node_modules
