+++ android/settings.gradle
@@ -1,7 +1,5 @@
-rootProject.name = 'mobileapp'
+rootProject.name = 'PostAR'
 include ':react-native-sensors'
 project(':react-native-sensors').projectDir = new File(rootProject.projectDir, '../node_modules/react-native-sensors/android')
-include ':react-native-camera'
-project(':react-native-camera').projectDir = new File(rootProject.projectDir, '../node_modules/react-native-camera/android')
 
 include ':app'
