+++ app.json
@@ -1,0 +1,4 @@
+{
+  "name": "PostAR",
+  "displayName": "PostAR"
+}
