+++ ios/PostARTests/PostARTests.m
@@ -16,11 +16,11 @@
 #define TIMEOUT_SECONDS 600
 #define TEXT_TO_LOOK_FOR @"Welcome to React Native!"
 
-@interface mobileappTests : XCTestCase
+@interface PostARTests : XCTestCase
 
 @end
 
-@implementation mobileappTests
+@implementation PostARTests
 
 - (BOOL)findSubviewInView:(UIView *)view matching:(BOOL(^)(UIView *view))test
 {
